#966: Reorganize inline methods
