+++ include/qengine_cpu.hpp
@@ -93,115 +93,13 @@ public:
 
     virtual void FreeStateVec(complex* sv = NULL) { stateVec = NULL; }
 
-    virtual void GetAmplitudePage(complex* pagePtr, bitCapIntOcl offset, bitCapIntOcl length)
-    {
-        Finish();
-
-        if (stateVec) {
-            stateVec->copy_out(pagePtr, offset, length);
-        } else {
-            std::fill(pagePtr, pagePtr + length, ZERO_CMPLX);
-        }
-    }
-    virtual void SetAmplitudePage(const complex* pagePtr, bitCapIntOcl offset, bitCapIntOcl length)
-    {
-        if (!stateVec) {
-            ResetStateVec(AllocStateVec(maxQPowerOcl));
-            stateVec->clear();
-        }
-
-        Finish();
-
-        stateVec->copy_in(pagePtr, offset, length);
-
-        runningNorm = REAL1_DEFAULT_ARG;
-    }
-    virtual void SetAmplitudePage(
-        QEnginePtr pageEnginePtr, bitCapIntOcl srcOffset, bitCapIntOcl dstOffset, bitCapIntOcl length)
-    {
-        QEngineCPUPtr pageEngineCpuPtr = std::dynamic_pointer_cast<QEngineCPU>(pageEnginePtr);
-        StateVectorPtr oStateVec = pageEngineCpuPtr->stateVec;
-
-        if (!stateVec && !oStateVec) {
-            return;
-        }
-
-        if (!oStateVec && (length == maxQPower)) {
-            ZeroAmplitudes();
-            return;
-        }
-
-        if (!stateVec) {
-            ResetStateVec(AllocStateVec(maxQPowerOcl));
-            stateVec->clear();
-        }
-
-        Finish();
-        pageEngineCpuPtr->Finish();
-
-        stateVec->copy_in(oStateVec, srcOffset, dstOffset, length);
-
-        runningNorm = REAL1_DEFAULT_ARG;
-    }
-    virtual void ShuffleBuffers(QEnginePtr engine)
-    {
-        QEngineCPUPtr engineCpu = std::dynamic_pointer_cast<QEngineCPU>(engine);
-
-        if (!stateVec && !(engineCpu->stateVec)) {
-            return;
-        }
-
-        if (!stateVec) {
-            ResetStateVec(AllocStateVec(maxQPowerOcl));
-            stateVec->clear();
-        }
-
-        if (!(engineCpu->stateVec)) {
-            engineCpu->ResetStateVec(engineCpu->AllocStateVec(maxQPowerOcl));
-            engineCpu->stateVec->clear();
-        }
-
-        Finish();
-        engineCpu->Finish();
-
-        stateVec->shuffle(engineCpu->stateVec);
-
-        runningNorm = REAL1_DEFAULT_ARG;
-        engineCpu->runningNorm = REAL1_DEFAULT_ARG;
-    }
-
     virtual bool IsZeroAmplitude() { return !stateVec; }
-
-    virtual void CopyStateVec(QEnginePtr src)
-    {
-        if (src->IsZeroAmplitude()) {
-            ZeroAmplitudes();
-            return;
-        }
-
-        if (!stateVec) {
-            ResetStateVec(AllocStateVec(maxQPowerOcl));
-        }
-
-        Finish();
-        src->Finish();
-
-        complex* sv;
-        if (isSparse) {
-            sv = new complex[(bitCapIntOcl)maxQPower];
-        } else {
-            sv = std::dynamic_pointer_cast<StateVectorArray>(stateVec)->amplitudes;
-        }
-
-        src->GetQuantumState(sv);
-
-        if (isSparse) {
-            SetQuantumState(sv);
-            delete[] sv;
-        }
-
-        runningNorm = src->GetRunningNorm();
-    }
+    virtual void GetAmplitudePage(complex* pagePtr, bitCapIntOcl offset, bitCapIntOcl length);
+    virtual void SetAmplitudePage(const complex* pagePtr, bitCapIntOcl offset, bitCapIntOcl length);
+    virtual void SetAmplitudePage(
+        QEnginePtr pageEnginePtr, bitCapIntOcl srcOffset, bitCapIntOcl dstOffset, bitCapIntOcl length);
+    virtual void ShuffleBuffers(QEnginePtr engine);
+    virtual void CopyStateVec(QEnginePtr src);
 
     virtual QEnginePtr CloneEmpty();
