+++ include/qinterface.hpp
@@ -176,8 +176,6 @@ protected:
         return toClamp;
     }
 
-    template <typename GateFunc> void ControlledLoopFixture(bitLenInt length, GateFunc gate);
-
     void FreeAligned(void* toFree)
     {
         if (toFree) {
