+++ include/qstabilizerhybrid.hpp
@@ -46,140 +46,19 @@ protected:
     QStabilizerPtr MakeStabilizer(bitCapInt perm = 0);
     QEnginePtr MakeEngine(bitCapInt perm = 0);
 
-    void InvertBuffer(bitLenInt qubit)
-    {
-        complex pauliX[4] = { ZERO_CMPLX, ONE_CMPLX, ONE_CMPLX, ZERO_CMPLX };
-        MpsShardPtr pauliShard = std::make_shared<MpsShard>(pauliX);
-        pauliShard->Compose(shards[qubit]->gate);
-        shards[qubit] = pauliShard->IsIdentity() ? NULL : pauliShard;
-        stabilizer->X(qubit);
-    }
-
-    void FlushIfBlocked(bitLenInt control, bitLenInt target, bool isPhase = false)
-    {
-        if (engine) {
-            return;
-        }
-
-        if (shards[target] && shards[target]->IsInvert()) {
-            InvertBuffer(target);
-        }
-
-        if (shards[control] && shards[control]->IsInvert()) {
-            InvertBuffer(control);
-        }
-
-        bool isBlocked = (shards[target] && (!isPhase || !shards[target]->IsPhase()));
-        isBlocked |= (shards[control] && !shards[control]->IsPhase());
-
-        if (isBlocked) {
-            SwitchToEngine();
-        }
-    }
-
-    virtual bool CollapseSeparableShard(bitLenInt qubit)
-    {
-        MpsShardPtr shard = shards[qubit];
-        shards[qubit] = NULL;
-        real1_f prob;
-
-        const bool isZ1 = stabilizer->M(qubit);
-
-        if (isZ1) {
-            prob = (real1_f)norm(shard->gate[3]);
-        } else {
-            prob = (real1_f)norm(shard->gate[2]);
-        }
-
-        bool result;
-        if (prob <= ZERO_R1) {
-            result = false;
-        } else if (prob >= ONE_R1) {
-            result = true;
-        } else {
-            result = (Rand() <= prob);
-        }
-
-        if (result != isZ1) {
-            stabilizer->X(qubit);
-        }
-
-        return result;
-    }
-
-    virtual void FlushBuffers()
-    {
-        if (stabilizer) {
-            for (bitLenInt i = 0; i < qubitCount; i++) {
-                if (shards[i]) {
-                    // This will call FlushBuffers() again after no longer stabilizer.
-                    SwitchToEngine();
-                    return;
-                }
-            }
-        }
-
-        if (stabilizer) {
-            return;
-        }
-
-        for (bitLenInt i = 0; i < qubitCount; i++) {
-            MpsShardPtr shard = shards[i];
-            if (shard) {
-                shards[i] = NULL;
-                engine->Mtrx(shard->gate, i);
-            }
-        }
-    }
-
+    virtual void InvertBuffer(bitLenInt qubit);
+    virtual void FlushIfBlocked(bitLenInt control, bitLenInt target, bool isPhase = false);
+    virtual bool CollapseSeparableShard(bitLenInt qubit);
+    virtual bool TrimControls(
+        const bitLenInt* lControls, bitLenInt lControlLen, std::vector<bitLenInt>& output, bool anti = false);
+    virtual void CacheEigenstate(bitLenInt target);
+    virtual void FlushBuffers();
     virtual void DumpBuffers()
     {
         for (bitLenInt i = 0; i < qubitCount; i++) {
             shards[i] = NULL;
         }
     }
-
-    virtual bool TrimControls(
-        const bitLenInt* lControls, bitLenInt lControlLen, std::vector<bitLenInt>& output, bool anti = false)
-    {
-        if (engine) {
-            output.insert(output.begin(), lControls, lControls + lControlLen);
-            return false;
-        }
-
-        for (bitLenInt i = 0; i < lControlLen; i++) {
-            bitLenInt bit = lControls[i];
-
-            if (!stabilizer->IsSeparableZ(bit)) {
-                output.push_back(bit);
-                continue;
-            }
-
-            if (shards[bit]) {
-                if (shards[bit]->IsInvert()) {
-                    if (anti != stabilizer->M(bit)) {
-                        return true;
-                    }
-                    continue;
-                }
-
-                if (shards[bit]->IsPhase()) {
-                    if (anti == stabilizer->M(bit)) {
-                        return true;
-                    }
-                    continue;
-                }
-
-                output.push_back(bit);
-            } else if (anti == stabilizer->M(bit)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    virtual void CacheEigenstate(bitLenInt target);
 
     virtual real1_f ApproxCompareHelper(
         QStabilizerHybridPtr toCompare, bool isDiscreteBool, real1_f error_tol = TRYDECOMPOSE_EPSILON);
@@ -985,23 +864,7 @@ public:
     }
 
     virtual void NormalizeState(
-        real1_f nrm = REAL1_DEFAULT_ARG, real1_f norm_thresh = REAL1_DEFAULT_ARG, real1_f phaseArg = ZERO_R1_F)
-    {
-        if (abs(nrm) <= FP_NORM_EPSILON) {
-            ZeroAmplitudes();
-            return;
-        }
-
-        if ((nrm > ZERO_R1) && (abs(ONE_R1 - nrm) > FP_NORM_EPSILON)) {
-            SwitchToEngine();
-        }
-
-        if (stabilizer) {
-            stabilizer->NormalizeState(REAL1_DEFAULT_ARG, norm_thresh, phaseArg);
-        } else {
-            engine->NormalizeState(nrm, norm_thresh, phaseArg);
-        }
-    }
+        real1_f nrm = REAL1_DEFAULT_ARG, real1_f norm_thresh = REAL1_DEFAULT_ARG, real1_f phaseArg = ZERO_R1_F);
 
     virtual real1_f ExpectationBitsAll(const bitLenInt* bits, bitLenInt length, bitCapInt offset = 0)
     {
@@ -1012,62 +875,9 @@ public:
         return engine->ExpectationBitsAll(bits, length, offset);
     }
 
-    virtual bool TrySeparate(bitLenInt qubit)
-    {
-        if (qubitCount == 1U) {
-            return true;
-        }
-
-        if (stabilizer) {
-            return stabilizer->CanDecomposeDispose(qubit, 1);
-        }
-
-        return engine->TrySeparate(qubit);
-    }
-    virtual bool TrySeparate(bitLenInt qubit1, bitLenInt qubit2)
-    {
-        if (qubitCount == 2U) {
-            return true;
-        }
-
-        if (stabilizer) {
-            if (qubit2 < qubit1) {
-                std::swap(qubit1, qubit2);
-            }
-
-            stabilizer->Swap(qubit1 + 1U, qubit2);
-
-            const bool toRet = stabilizer->CanDecomposeDispose(qubit1, 2);
-
-            stabilizer->Swap(qubit1 + 1U, qubit2);
-
-            return toRet;
-        }
-
-        return engine->TrySeparate(qubit1, qubit2);
-    }
-    virtual bool TrySeparate(const bitLenInt* qubits, bitLenInt length, real1_f error_tol)
-    {
-        if (stabilizer) {
-            std::vector<bitLenInt> q(length);
-            std::copy(qubits, qubits + length, q.begin());
-            std::sort(q.begin(), q.end());
-
-            for (bitLenInt i = 1; i < length; i++) {
-                Swap(q[0] + i, q[i]);
-            }
-
-            const bool toRet = stabilizer->CanDecomposeDispose(q[0], length);
-
-            for (bitLenInt i = 1; i < length; i++) {
-                Swap(q[0] + i, q[i]);
-            }
-
-            return toRet;
-        }
-
-        return engine->TrySeparate(qubits, length, error_tol);
-    }
+    virtual bool TrySeparate(bitLenInt qubit);
+    virtual bool TrySeparate(bitLenInt qubit1, bitLenInt qubit2);
+    virtual bool TrySeparate(const bitLenInt* qubits, bitLenInt length, real1_f error_tol);
 
     virtual QInterfacePtr Clone();
