+++ src/qpager.cpp
@@ -191,6 +191,28 @@ QEnginePtr QPager::MakeEngine(bitLenInt length, bitCapInt perm, int deviceId)
         false, false, useHostRam, deviceId, useRDRAND, isSparse, (real1_f)amplitudeFloor));
 }
 
+void QPager::GetSetAmplitudePage(complex* pagePtr, const complex* cPagePtr, bitCapIntOcl offset, bitCapIntOcl length)
+{
+    const bitCapIntOcl pageLength = (bitCapIntOcl)pageMaxQPower();
+    bitCapIntOcl perm = 0U;
+    for (bitCapIntOcl i = 0; i < qPages.size(); i++) {
+        if ((perm + length) < offset) {
+            continue;
+        }
+        if (perm >= (offset + length)) {
+            break;
+        }
+        const bitCapInt partOffset = (perm < offset) ? (offset - perm) : 0U;
+        const bitCapInt partLength = (length < pageLength) ? length : pageLength;
+        if (cPagePtr) {
+            qPages[i]->SetAmplitudePage(cPagePtr, (bitCapIntOcl)partOffset, (bitCapIntOcl)partLength);
+        } else {
+            qPages[i]->GetAmplitudePage(pagePtr, (bitCapIntOcl)partOffset, (bitCapIntOcl)partLength);
+        }
+        perm += pageLength;
+    }
+}
+
 void QPager::CombineEngines(bitLenInt bit)
 {
     if (bit > qubitCount) {
